search/genai-101/code_examples/chat-app-code_es-8-15/requirements.txt
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.1
Pygments==2.18.0
annotated-types==0.7.0
anyio==4.4.0
boto3==1.34.136
botocore==1.34.136
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
ecs-logging==2.2.0
eland==8.15.0
eland[pytorch]==8.15.0
elastic-apm==6.22.3
elastic-transport==8.13.1
elasticsearch==8.15.0
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi-cli==0.0.4
fastapi==0.111.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
orjson==3.10.5
psutil==6.0.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
torch==2.1.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.15
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
